+++ app/Http/Controllers/DanhMucController.php
@@ -6,12 +6,26 @@ use App\Http\Requests\CapNhatDanhMucRequest;
 use App\Http\Requests\ThemMoiDanhMucRequest;
 use App\Http\Requests\XoaDanhMucRequest;
 use App\Models\DanhMuc;
+use App\Models\PhanQuyen;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class DanhMucController extends Controller
 {
     public function getData()
     {
+        $id_chuc_nang = 22;
+        $login = Auth::guard('sanctum')->user();
+        $id_chuc_vu = $login->id_chuc_vu;
+        $check_quyen = PhanQuyen::where('id_chuc_vu', $id_chuc_vu)
+                                ->where('id_chuc_nang', $id_chuc_nang)
+                                ->first();
+        if (!$check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $data = DanhMuc::join('danh_mucs as A', 'A.id', 'danh_mucs.id_danh_muc_cha')
                         ->select('danh_mucs.*', 'A.ten_danh_muc as ten_danh_muc_cha')
                         ->get();
@@ -22,6 +36,18 @@ class DanhMucController extends Controller
     }
     public function store(ThemMoiDanhMucRequest $request)
     {
+        $id_chuc_nang = 23;
+        $login = Auth::guard('sanctum')->user();
+        $id_chuc_vu = $login->id_chuc_vu;
+        $check_quyen = PhanQuyen::where('id_chuc_vu', $id_chuc_vu)
+                                ->where('id_chuc_nang', $id_chuc_nang)
+                                ->first();
+        if (!$check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $data = DanhMuc::create($request->all());
         return response()->json([
             'status' => 1,
@@ -30,6 +56,18 @@ class DanhMucController extends Controller
     }
     public function destroy(XoaDanhMucRequest $request)
     {
+        $id_chuc_nang = 25;
+        $login = Auth::guard('sanctum')->user();
+        $id_chuc_vu = $login->id_chuc_vu;
+        $check_quyen = PhanQuyen::where('id_chuc_vu', $id_chuc_vu)
+                                ->where('id_chuc_nang', $id_chuc_nang)
+                                ->first();
+        if (!$check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $data = DanhMuc::find($request->id);
         $data->delete();
         return response()->json([
@@ -39,6 +77,18 @@ class DanhMucController extends Controller
     }
     public function update(CapNhatDanhMucRequest $request)
     {
+        $id_chuc_nang = 24;
+        $login = Auth::guard('sanctum')->user();
+        $id_chuc_vu = $login->id_chuc_vu;
+        $check_quyen = PhanQuyen::where('id_chuc_vu', $id_chuc_vu)
+                                ->where('id_chuc_nang', $id_chuc_nang)
+                                ->first();
+        if (!$check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $data = DanhMuc::find($request->id);
         $data->update($request->all());
         return response()->json([
@@ -48,6 +98,18 @@ class DanhMucController extends Controller
     }
     public function changeStatus(XoaDanhMucRequest $request)
     {
+        $id_chuc_nang = 26;
+        $login = Auth::guard('sanctum')->user();
+        $id_chuc_vu = $login->id_chuc_vu;
+        $check_quyen = PhanQuyen::where('id_chuc_vu', $id_chuc_vu)
+                                ->where('id_chuc_nang', $id_chuc_nang)
+                                ->first();
+        if (!$check_quyen) {
+            return response()->json([
+                'data' => false,
+                'message' => "bạn không có quyền thực hiện chức năng này!"
+            ]);
+        }
         $data = DanhMuc::find($request->id);
         if ($data->tinh_trang == 1) {
             $data->tinh_trang = 0;
